+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 2025-06-25 08:07:31 - Bump version to $(node -p 'require(\'./package.json\').version')
+
+### Changed Files:
+- `.github/workflows/release.yml`
+- `custom-fields-block.php`
+
 ## 2025-06-25 08:01:57 - 1.2.7
 
 ### Changed Files: